+++ packages/rs-dpp/src/data_contract/state_transition/data_contract_update_transition/validation/basic/validate_indices_are_backward_compatible.rs
@@ -1,6 +1,5 @@
 use std::collections::HashSet;
 use std::collections::{BTreeMap, HashMap};
-use std::iter::FromIterator;
 
 use crate::consensus::basic::BasicError;
 use crate::data_contract::document_type::IndexProperty;
