+++ packages/rs-dpp/src/data_contract/validation/data_contract_validator.rs
@@ -1,4 +1,3 @@
-use std::collections::BTreeMap;
 use std::{collections::HashMap, sync::Arc};
 
 use anyhow::anyhow;
