+++ packages/rs-drive/src/drive/mod.rs
@@ -202,25 +202,6 @@ impl Drive {
             .map_err(Error::GroveDB)
     }
 
-    /// Make sure the protocol version is correct.
-    pub const fn check_protocol_version(_version: u32) -> bool {
-        // Temporary disabled due protocol version is dynamic and goes from consensus params
-        true
-    }
-
-    /// Makes sure the protocol version is correct given the version as a u8.
-    pub fn check_protocol_version_bytes(version_bytes: &[u8]) -> bool {
-        if version_bytes.len() != 4 {
-            false
-        } else {
-            let version_set_bytes: [u8; 4] = version_bytes
-                .try_into()
-                .expect("slice with incorrect length");
-            let version = u32::from_be_bytes(version_set_bytes);
-            Drive::check_protocol_version(version)
-        }
-    }
-
     /// Applies a batch of Drive operations to groveDB.
     fn apply_batch_drive_operations(
         &self,
@@ -304,27 +285,5 @@ impl Drive {
             false,
             None,
         )
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::collections::HashMap;
-    use std::option::Option::None;
-
-    use dpp::data_contract::extra::common::json_document_to_cbor;
-    use tempfile::TempDir;
-
-    use crate::drive::Drive;
-
-    #[test]
-    fn test_cbor_deserialization() {
-        let serialized_document =
-            json_document_to_cbor("simple.json", Some(1)).expect("expected to get cbor contract");
-        let (version, read_serialized_document) = serialized_document.split_at(4);
-        assert!(Drive::check_protocol_version_bytes(version));
-        let document: HashMap<String, ciborium::value::Value> =
-            ciborium::de::from_reader(read_serialized_document).expect("cannot deserialize cbor");
-        assert!(document.get("a").is_some());
     }
 }
